+++ Tests/StringCommentingTests.swift
@@ -9,7 +9,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2019-2023 1024jp
+//  © 2019-2024 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -33,70 +33,49 @@ final class StringCommentingTests: XCTestCase {
     
     func testInlineCommentOut() {
         
-        XCTAssertEqual("foo".inlineCommentOut(delimiter: "//", spacer: "", ranges: []), [])
+        XCTAssertEqual("foo".inlineCommentOut(delimiter: "//", ranges: []), [])
         
-        XCTAssertEqual("foo".inlineCommentOut(delimiter: "//", spacer: "", ranges: [NSRange(0..<0)]),
+        XCTAssertEqual("foo".inlineCommentOut(delimiter: "//", ranges: [NSRange(0..<0)]),
                        [.init(string: "//", location: 0, forward: true)])
-        XCTAssertEqual("foo".inlineCommentOut(delimiter: "//", spacer: " ", ranges: [NSRange(0..<0)]),
-                       [.init(string: "// ", location: 0, forward: true)])
-        XCTAssertEqual("foo".inlineCommentOut(delimiter: "//", spacer: "", ranges: [NSRange(1..<2)]),
+        XCTAssertEqual("foo".inlineCommentOut(delimiter: "//", ranges: [NSRange(1..<2)]),
                        [.init(string: "//", location: 1, forward: true)])
     }
     
     
     func testBlockCommentOut() {
         
-        XCTAssertEqual("foo".blockCommentOut(delimiters: Pair("<-", "->"), spacer: "", ranges: []), [])
+        XCTAssertEqual("foo".blockCommentOut(delimiters: Pair("<-", "->"), ranges: []), [])
         
-        XCTAssertEqual("foo".blockCommentOut(delimiters: Pair("<-", "->"), spacer: "", ranges: [NSRange(0..<0)]),
+        XCTAssertEqual("foo".blockCommentOut(delimiters: Pair("<-", "->"), ranges: [NSRange(0..<0)]),
                        [.init(string: "<-", location: 0, forward: true), .init(string: "->", location: 0, forward: false)])
-        XCTAssertEqual("foo".blockCommentOut(delimiters: Pair("<-", "->"), spacer: " ", ranges: [NSRange(0..<3)]),
-                       [.init(string: "<- ", location: 0, forward: true), .init(string: " ->", location: 3, forward: false)])
-        XCTAssertEqual("foo".blockCommentOut(delimiters: Pair("<-", "->"), spacer: " ", ranges: [NSRange(1..<2)]),
-                       [.init(string: "<- ", location: 1, forward: true), .init(string: " ->", location: 2, forward: false)])
     }
     
     
     func testInlineUncomment() {
         
-        XCTAssertEqual("foo".rangesOfInlineDelimiter("//", spacer: "", ranges: []), [])
-        XCTAssertEqual("foo".rangesOfInlineDelimiter("//", spacer: "", ranges: [NSRange(0..<0)]), [])
+        XCTAssertEqual("foo".rangesOfInlineDelimiter("//", ranges: []), [])
+        XCTAssertEqual("foo".rangesOfInlineDelimiter("//", ranges: [NSRange(0..<0)]), [])
         
-        XCTAssertEqual("//foo".rangesOfInlineDelimiter("//", spacer: "", ranges: [NSRange(0..<5)]), [NSRange(0..<2)])
-        XCTAssertEqual("//foo".rangesOfInlineDelimiter("//", spacer: " ", ranges: [NSRange(0..<5)]), [NSRange(0..<2)])
-        XCTAssertEqual("// foo".rangesOfInlineDelimiter("//", spacer: "", ranges: [NSRange(0..<5)]), [NSRange(0..<2)])
-        XCTAssertEqual("// foo".rangesOfInlineDelimiter("//", spacer: " ", ranges: [NSRange(0..<5)]), [NSRange(0..<3)])
+        XCTAssertEqual("//foo".rangesOfInlineDelimiter("//", ranges: [NSRange(0..<5)]), [NSRange(0..<2)])
+        XCTAssertEqual("// foo".rangesOfInlineDelimiter("//", ranges: [NSRange(0..<5)]), [NSRange(0..<2)])
         
-        XCTAssertEqual("  //foo".rangesOfInlineDelimiter("//", spacer: "", ranges: [NSRange(0..<7)]), [NSRange(2..<4)])
-        XCTAssertNil("// foo".rangesOfInlineDelimiter("//", spacer: " ", ranges: [NSRange(0..<1)]))
-        XCTAssertNil("// foo".rangesOfInlineDelimiter("//", spacer: " ", ranges: [NSRange(1..<3)]))
-        XCTAssertEqual("// foo".rangesOfInlineDelimiter("//", spacer: " ", ranges: [NSRange(0..<2)]), [NSRange(0..<2)])
-        
-        XCTAssertEqual("// foo\n//bar".rangesOfInlineDelimiter("//", spacer: " ", ranges: [NSRange(0..<12)]), [NSRange(0..<3), NSRange(7..<9)])
-        XCTAssertEqual(" //foo\n//bar".rangesOfInlineDelimiter("//", spacer: " ", ranges: [NSRange(0..<12)]), [NSRange(1..<3), NSRange(7..<9)])
-        XCTAssertEqual(" //foo\n//bar".rangesOfInlineDelimiter("//", spacer: " ", ranges: [NSRange(0..<3), NSRange(0..<6)]), [NSRange(1..<3)])
-        
-        XCTAssertEqual("// foo\n//bar".rangesOfInlineDelimiter("//", spacer: " ", ranges: [NSRange(0..<5), NSRange(7..<12)]), [NSRange(0..<3), NSRange(7..<9)])
+        XCTAssertEqual("  //foo".rangesOfInlineDelimiter("//", ranges: [NSRange(0..<7)]), [NSRange(2..<4)])
     }
     
     
     func testBlockUncomment() {
         
-        XCTAssertEqual("foo".rangesOfBlockDelimiters(Pair("<-", "->"), spacer: "", ranges: []), [])
-        XCTAssertEqual("foo".rangesOfBlockDelimiters(Pair("<-", "->"), spacer: "", ranges: [NSRange(0..<0)]), [])
+        XCTAssertEqual("foo".rangesOfBlockDelimiters(Pair("<-", "->"), ranges: []), [])
+        XCTAssertEqual("foo".rangesOfBlockDelimiters(Pair("<-", "->"), ranges: [NSRange(0..<0)]), [])
         
-        XCTAssertEqual("<-foo->".rangesOfBlockDelimiters(Pair("<-", "->"), spacer: "", ranges: [NSRange(0..<7)]), [NSRange(0..<2), NSRange(5..<7)])
-        XCTAssertEqual("<-foo->".rangesOfBlockDelimiters(Pair("<-", "->"), spacer: " ", ranges: [NSRange(0..<7)]), [NSRange(0..<2), NSRange(5..<7)])
-        XCTAssertEqual("<- foo ->".rangesOfBlockDelimiters(Pair("<-", "->"), spacer: "", ranges: [NSRange(0..<9)]), [NSRange(0..<2), NSRange(7..<9)])
-        XCTAssertEqual("<- foo ->".rangesOfBlockDelimiters(Pair("<-", "->"), spacer: " ", ranges: [NSRange(0..<9)]), [NSRange(0..<3), NSRange(6..<9)])
+        XCTAssertEqual("<-foo->".rangesOfBlockDelimiters(Pair("<-", "->"), ranges: [NSRange(0..<7)]), [NSRange(0..<2), NSRange(5..<7)])
+        XCTAssertEqual("<- foo ->".rangesOfBlockDelimiters(Pair("<-", "->"), ranges: [NSRange(0..<9)]), [NSRange(0..<2), NSRange(7..<9)])
         
-        XCTAssertEqual(" <-foo-> ".rangesOfBlockDelimiters(Pair("<-", "->"), spacer: "", ranges: [NSRange(0..<9)]), [NSRange(1..<3), NSRange(6..<8)])
-        XCTAssertNil(" <-foo-> ".rangesOfBlockDelimiters(Pair("<-", "->"), spacer: "", ranges: [NSRange(1..<7)]))
+        XCTAssertEqual(" <-foo-> ".rangesOfBlockDelimiters(Pair("<-", "->"), ranges: [NSRange(0..<9)]), [NSRange(1..<3), NSRange(6..<8)])
+        XCTAssertNil(" <-foo-> ".rangesOfBlockDelimiters(Pair("<-", "->"), ranges: [NSRange(1..<7)]))
         
         // ok, this is currently in spec, but not a good one...
-        XCTAssertEqual("<-foo-><-bar->".rangesOfBlockDelimiters(Pair("<-", "->"), spacer: "", ranges: [NSRange(0..<14)]), [NSRange(0..<2), NSRange(12..<14)])
-        
-        XCTAssertNil("<- foo ->".rangesOfBlockDelimiters(Pair("<-", "->"), spacer: " ", ranges: [NSRange(0..<3), NSRange(6..<9)]))
+        XCTAssertEqual("<-foo-><-bar->".rangesOfBlockDelimiters(Pair("<-", "->"), ranges: [NSRange(0..<14)]), [NSRange(0..<2), NSRange(12..<14)])
     }
     
     
@@ -110,8 +89,8 @@ final class StringCommentingTests: XCTestCase {
         textView.string = "foo\nbar"
         textView.selectedRanges = [NSRange(0..<3), NSRange(4..<7)] as [NSValue]
         textView.commentOut(types: .inline, fromLineHead: true)
-        XCTAssertEqual(textView.string, "// foo\n// bar")
-        XCTAssertEqual(textView.selectedRanges, [NSRange(0..<6), NSRange(7..<13)] as [NSValue])
+        XCTAssertEqual(textView.string, "//foo\n//bar")
+        XCTAssertEqual(textView.selectedRanges, [NSRange(0..<5), NSRange(6..<11)] as [NSValue])
         XCTAssertTrue(textView.canUncomment(partly: false))
         textView.uncomment()
         XCTAssertEqual(textView.string, "foo\nbar")
@@ -120,8 +99,8 @@ final class StringCommentingTests: XCTestCase {
         textView.selectedRanges = [NSRange(1..<1)] as [NSValue]
         textView.insertionLocations = [5]
         textView.commentOut(types: .inline, fromLineHead: true)
-        XCTAssertEqual(textView.string, "// foo\n// bar")
-        XCTAssertEqual(textView.rangesForUserTextChange, [NSRange(4..<4), NSRange(11..<11)] as [NSValue])
+        XCTAssertEqual(textView.string, "//foo\n//bar")
+        XCTAssertEqual(textView.rangesForUserTextChange, [NSRange(3..<3), NSRange(9..<9)] as [NSValue])
         XCTAssertTrue(textView.canUncomment(partly: false))
         textView.uncomment()
         XCTAssertEqual(textView.string, "foo\nbar")
@@ -136,8 +115,8 @@ final class StringCommentingTests: XCTestCase {
         textView.string = "foo\nbar"
         textView.selectedRanges = [NSRange(0..<3), NSRange(4..<7)] as [NSValue]
         textView.commentOut(types: .block, fromLineHead: true)
-        XCTAssertEqual(textView.string, "<- foo ->\n<- bar ->")
-        XCTAssertEqual(textView.selectedRanges, [NSRange(0..<9), NSRange(10..<19)] as [NSValue])
+        XCTAssertEqual(textView.string, "<-foo->\n<-bar->")
+        XCTAssertEqual(textView.selectedRanges, [NSRange(0..<7), NSRange(8..<15)] as [NSValue])
         XCTAssertTrue(textView.canUncomment(partly: false))
         textView.uncomment()
         XCTAssertEqual(textView.string, "foo\nbar")
@@ -146,8 +125,8 @@ final class StringCommentingTests: XCTestCase {
         textView.selectedRanges = [NSRange(1..<1)] as [NSValue]
         textView.insertionLocations = [5]
         textView.commentOut(types: .block, fromLineHead: true)
-        XCTAssertEqual(textView.string, "<- foo ->\n<- bar ->")
-        XCTAssertEqual(textView.rangesForUserTextChange, [NSRange(4..<4), NSRange(14..<14)] as [NSValue])
+        XCTAssertEqual(textView.string, "<-foo->\n<-bar->")
+        XCTAssertEqual(textView.rangesForUserTextChange, [NSRange(3..<3), NSRange(11..<11)] as [NSValue])
         XCTAssertTrue(textView.canUncomment(partly: false))
         textView.uncomment()
         XCTAssertEqual(textView.string, "foo\nbar")
@@ -186,7 +165,6 @@ private final class CommentingTextView: NSTextView, Commenting, MultiCursorEditi
     // Commenting
     var inlineCommentDelimiter: String? = "//"
     var blockCommentDelimiters: Pair<String>? = Pair("<-", "->")
-    var appendsCommentSpacer: Bool = true
     
     // MultiCursorEditing
     var insertionLocations: [Int] = []