+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Improve the look and feel of the syntax editor.
 - Make the encoding priority view undoable.
+- Deprecate “Append a space to comment delimiter” option
 - [trivial] Enable delete selected separators in the encoding priority view with the Delete key.
 - [trivial] Avoid users using reserved syntax names.
 - [dev] Migrate the settings window and syntax editor to SwiftUI.